file restructure for separate css styling
