+++ client/src/components/UserDashboard/UserDashboard.jsx
@@ -5,6 +5,7 @@ import MenuItem from '@material-ui/core/MenuItem';
 import FormHelperText from '@material-ui/core/FormHelperText';
 import FormControl from '@material-ui/core/FormControl';
 import Select from '@material-ui/core/Select';
+import './UserDashboard.css';
 
 //PAGE ELEMENTS TO INCLUDE
 // Player Avatar
@@ -13,18 +14,18 @@ import Select from '@material-ui/core/Select';
 // Player Tournament History (Tournament Name, Game Name, W or L)
 // Nearby/Upcoming Tournaments Viewer
 
-const useStyles = makeStyles((theme) => ({
-  formControl: {
-    margin: theme.spacing(1),
-    minWidth: 120,
-  },
-  selectEmpty: {
-    marginTop: theme.spacing(2),
-  },
-}));
+// const useStyles = makeStyles((theme) => ({
+//   formControl: {
+//     margin: theme.spacing(1),
+//     minWidth: 120,
+//   },
+//   selectEmpty: {
+//     marginTop: theme.spacing(2),
+//   },
+// }));
 
 const UserDashboard = (props) => {
-  const classes = useStyles();
+  // const classes = useStyles();
   const [tournamentStyle, setTournamentStyle] = useState('');
 
   const handleTournamentStyleChange = (event) => {
@@ -33,8 +34,8 @@ const UserDashboard = (props) => {
 
   return (
     <div>
-      <FormControl variant="outlined" className={classes.formControl}>
-        <InputLabel id="tournament style selector">Select Tournament Style</InputLabel>
+      <FormControl variant="outlined" className="formControl">
+        <InputLabel id="tournament style selector">Style</InputLabel>
         <Select
           labelId="tournament style"
           id="tournament style"
